pyproject.toml
[tool.poetry]
name = "clientes"
version = "0.1.0"
description = ""
authors = ["José Guilherme <josegfernandes03@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
asgiref = "^3.6.0"
astroid = "^2.15.0"
django = "^4.1.7"
djangorestframework = "^3.14.0"
isort = "^5.12.0"
lazy-object-proxy = "^1.9.0"
mccabe = "^0.7.0"
pylint = "^2.16.4"
pytz = "^2022.7.1"
six = "^1.16.0"
sqlparse = "^0.4.3"
toml = "^0.10.2"
wrapt = "^1.15.0"
mypy = "^1.0.1"
python-dotenv = "^1.0.0"
validate-docbr = "^1.10.0"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
